requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
flask==3.0.3
importlib-metadata==8.2.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
PuLP==2.9.0
werkzeug==3.0.3
zipp==3.19.2
ortools==9.11.4210
numpy==1.24.3
openai==0.27.0
flask-cors==4.0.0
python-dotenv==1.0.1
